@@ -1,58 +1,62 @@
-import * as fs from 'fs';
-import * as path from 'path';
+import * as fs from 'fs-extra';
+import 'reflect-metadata';
+import { EnvironmentVariablesService } from './variables/environment';
+import { EnvironmentVariables } from './variables/types';
+export const IS_WINDOWS = /^win/.test(process.platform);
 
-type EnvVars = Object & { [key: string]: string };
+function parseEnvironmentVariables(contents: string): EnvironmentVariables | undefined {
+    if (typeof contents !== 'string' || contents.length === 0) {
+        return undefined;
+    }
 
-export function parseEnvFile(envFile: string, mergeWithProcessEnvVars: boolean = true): EnvVars {
-    const buffer = fs.readFileSync(envFile, 'utf8');
-    const env = {};
-    buffer.split('\n').forEach(line => {
-        const r = line.match(/^\s*([\w\.\-]+)\s*=\s*(.*)?\s*$/);
-        if (r !== null) {
-            let value = r[2] || '';
+    const env = {} as EnvironmentVariables;
+    contents.split('\n').forEach(line => {
+        const match = line.match(/^\s*([\w\.\-]+)\s*=\s*(.*)?\s*$/);
+        if (match !== null) {
+            let value = typeof match[2] === 'string' ? match[2] : '';
             if (value.length > 0 && value.charAt(0) === '"' && value.charAt(value.length - 1) === '"') {
                 value = value.replace(/\\n/gm, '\n');
             }
-            env[r[1]] = value.replace(/(^['"]|['"]$)/g, '');
+            env[match[1]] = value.replace(/(^['"]|['"]$)/g, '');
         }
     });
-    return mergeWithProcessEnvVars ? mergeEnvVariables(env, process.env) : mergePythonPath(env, process.env.PYTHONPATH);
+    return env;
+}
+
+export function parseEnvFile(envFile: string, mergeWithProcessEnvVars: boolean = true): EnvironmentVariables {
+    const buffer = fs.readFileSync(envFile, 'utf8');
+    const env = parseEnvironmentVariables(buffer)!;
+    return mergeWithProcessEnvVars ? mergeEnvVariables(env, process.env) : mergePythonPath(env, process.env.PYTHONPATH as string);
 }
 
 /**
  * Merge the target environment variables into the source.
  * Note: The source variables are modified and returned (i.e. it modifies value passed in).
  * @export
- * @param {EnvVars} targetEnvVars target environment variables.
- * @param {EnvVars} [sourceEnvVars=process.env] source environment variables (defaults to current process variables).
- * @returns {EnvVars}
+ * @param {EnvironmentVariables} targetEnvVars target environment variables.
+ * @param {EnvironmentVariables} [sourceEnvVars=process.env] source environment variables (defaults to current process variables).
+ * @returns {EnvironmentVariables}
  */
-export function mergeEnvVariables(targetEnvVars: EnvVars, sourceEnvVars: EnvVars = process.env): EnvVars {
-    Object.keys(sourceEnvVars).forEach(setting => {
-        if (targetEnvVars[setting] === undefined) {
-            targetEnvVars[setting] = sourceEnvVars[setting];
-        }
-    });
-    return mergePythonPath(targetEnvVars, sourceEnvVars.PYTHONPATH);
+export function mergeEnvVariables(targetEnvVars: EnvironmentVariables, sourceEnvVars: EnvironmentVariables = process.env): EnvironmentVariables {
+    const service = new EnvironmentVariablesService(IS_WINDOWS);
+    service.mergeVariables(sourceEnvVars, targetEnvVars);
+    service.appendPythonPath(targetEnvVars, sourceEnvVars.PYTHONPATH);
+    return targetEnvVars;
 }
 
 /**
  * Merge the target PYTHONPATH value into the env variables passed.
  * Note: The env variables passed in are modified and returned (i.e. it modifies value passed in).
  * @export
- * @param {EnvVars} env target environment variables.
+ * @param {EnvironmentVariables} env target environment variables.
  * @param {string | undefined} [currentPythonPath] PYTHONPATH value.
- * @returns {EnvVars}
+ * @returns {EnvironmentVariables}
  */
-export function mergePythonPath(env: EnvVars, currentPythonPath: string | undefined): EnvVars {
+export function mergePythonPath(env: EnvironmentVariables, currentPythonPath: string | undefined): EnvironmentVariables {
     if (typeof currentPythonPath !== 'string' || currentPythonPath.length === 0) {
         return env;
     }
-
-    if (typeof env.PYTHONPATH === 'string' && env.PYTHONPATH.length > 0) {
-        env.PYTHONPATH = env.PYTHONPATH + path.delimiter + currentPythonPath;
-    } else {
-        env.PYTHONPATH = currentPythonPath;
-    }
+    const service = new EnvironmentVariablesService(IS_WINDOWS);
+    service.appendPythonPath(env, currentPythonPath!);
     return env;
 }
