@@ -25,14 +25,12 @@ export class Linter extends baseLinter.BaseLinter {
             pep8Args = ProductExecutableAndArgs.get(Product.pep8).args;
         }
 
-        return new Promise<baseLinter.ILintMessage[]>(resolve => {
-            this.run(pep8Path, pep8Args.concat(['--format=%(row)d,%(col)d,%(code).1s,%(code)s:%(text)s', document.uri.fsPath]), document, this.getWorkspaceRootPath(document), cancellation).then(messages => {
-                messages.forEach(msg => {
-                    msg.severity = this.parseMessagesSeverity(msg.type, this.pythonSettings.linting.pep8CategorySeverity);
-                });
-
-                resolve(messages);
+        return this.run(pep8Path, pep8Args.concat(['--format=%(row)d,%(col)d,%(code).1s,%(code)s:%(text)s', document.uri.fsPath]), document, this.getWorkspaceRootPath(document), cancellation).then(messages => {
+            messages.forEach(msg => {
+                msg.severity = this.parseMessagesSeverity(msg.type, this.pythonSettings.linting.pep8CategorySeverity);
             });
+
+            return messages;
         });
     }
 }
