@@ -5,11 +5,7 @@
 
 export class StopWatch {
     private started: number = Date.now();
-    private stopped?: number;
     public get elapsedTime() {
-        return (this.stopped ? this.stopped : Date.now()) - this.started;
-    }
-    public stop() {
-        this.stopped = Date.now();
+        return Date.now() - this.started;
     }
 }
