@@ -1,26 +1,37 @@
 'use strict';
-import * as vscode from 'vscode';
+import { Uri } from 'vscode';
 import { Product } from '../../common/installer';
-import { BaseTestManager } from '../common/baseTestManager';
-import { ITestCollectionStorageService, ITestDebugLauncher, ITestResultsService, ITestsHelper, Tests, TestsToRun } from '../common/types';
-import { discoverTests } from './collector';
+import { IServiceContainer } from '../../ioc/types';
+import { BaseTestManager } from '../common/managers/baseTestManager';
+import { TestDiscoveryOptions, TestRunOptions, Tests, TestsToRun } from '../common/types';
 import { runTest } from './runner';
 
 export class TestManager extends BaseTestManager {
-    constructor(rootDirectory: string, outputChannel: vscode.OutputChannel,
-        testCollectionStorage: ITestCollectionStorageService,
-        testResultsService: ITestResultsService, testsHelper: ITestsHelper, private debugLauncher: ITestDebugLauncher) {
-        super('pytest', Product.pytest, rootDirectory, outputChannel, testCollectionStorage, testResultsService, testsHelper);
+    constructor(workspaceFolder: Uri, rootDirectory: string,
+        serviceContainer: IServiceContainer) {
+        super('pytest', Product.pytest, workspaceFolder, rootDirectory, serviceContainer);
     }
-    public discoverTestsImpl(ignoreCache: boolean): Promise<Tests> {
+    public getDiscoveryOptions(ignoreCache: boolean): TestDiscoveryOptions {
         const args = this.settings.unitTest.pyTestArgs.slice(0);
-        return discoverTests(this.rootDirectory, args, this.testDiscoveryCancellationToken, ignoreCache, this.outputChannel, this.testsHelper);
+        return {
+            workspaceFolder: this.workspaceFolder,
+            cwd: this.rootDirectory, args,
+            token: this.testDiscoveryCancellationToken!, ignoreCache,
+            outChannel: this.outputChannel
+        };
     }
-    public runTestImpl(tests: Tests, testsToRun?: TestsToRun, runFailedTests?: boolean, debug?: boolean): Promise<{}> {
+    public async runTestImpl(tests: Tests, testsToRun?: TestsToRun, runFailedTests?: boolean, debug?: boolean): Promise<{}> {
         const args = this.settings.unitTest.pyTestArgs.slice(0);
         if (runFailedTests === true && args.indexOf('--lf') === -1 && args.indexOf('--last-failed') === -1) {
             args.push('--last-failed');
         }
-        return runTest(this.testResultsService, this.debugLauncher, this.rootDirectory, tests, args, testsToRun, this.testRunnerCancellationToken, this.outputChannel, debug);
+        const options: TestRunOptions = {
+            workspaceFolder: this.workspaceFolder,
+            cwd: this.rootDirectory,
+            tests, args, testsToRun, debug,
+            token: this.testRunnerCancellationToken!,
+            outChannel: this.outputChannel
+        };
+        return runTest(this.serviceContainer, this.testResultsService, options);
     }
 }
