@@ -1,14 +1,13 @@
 'use strict';
 import * as path from 'path';
-import { CancellationToken, OutputChannel, Uri } from 'vscode';
-import { PythonSettings } from '../../common/configSettings';
-import { BaseTestManager } from '../common/baseTestManager';
-import { run } from '../common/runner';
-import { ITestDebugLauncher, ITestResultsService, Tests, TestStatus, TestsToRun } from '../common/types';
-import { Server } from './socketServer';
+import { IServiceContainer } from '../../ioc/types';
+import { BaseTestManager } from '../common/managers/baseTestManager';
+import { Options, run } from '../common/runner';
+import { ITestDebugLauncher, ITestResultsService, IUnitTestSocketServer, TestRunOptions, Tests, TestStatus, TestsToRun } from '../common/types';
+
 type TestStatusMap = {
     status: TestStatus;
-    summaryProperty: string;
+    summaryProperty: 'passed' | 'failures' | 'errors' | 'skipped';
 };
 
 const outcomeMapping = new Map<string, TestStatusMap>();
@@ -29,14 +28,14 @@ interface ITestData {
 }
 
 // tslint:disable-next-line:max-func-body-length
-export function runTest(testManager: BaseTestManager, testResultsService: ITestResultsService, debugLauncher: ITestDebugLauncher, rootDirectory: string, tests: Tests, args: string[], testsToRun?: TestsToRun, token?: CancellationToken, outChannel?: OutputChannel, debug?: boolean): Promise<Tests> {
-    tests.summary.errors = 0;
-    tests.summary.failures = 0;
-    tests.summary.passed = 0;
-    tests.summary.skipped = 0;
+export function runTest(serviceContainer: IServiceContainer, testManager: BaseTestManager, testResultsService: ITestResultsService, options: TestRunOptions): Promise<Tests> {
+    options.tests.summary.errors = 0;
+    options.tests.summary.failures = 0;
+    options.tests.summary.passed = 0;
+    options.tests.summary.skipped = 0;
     let failFast = false;
     const testLauncherFile = path.join(__dirname, '..', '..', '..', '..', 'pythonFiles', 'PythonTools', 'visualstudio_py_testlauncher.py');
-    const server = new Server();
+    const server = serviceContainer.get<IUnitTestSocketServer>(IUnitTestSocketServer);
     server.on('error', (message: string, ...data: string[]) => {
         // tslint:disable-next-line:no-console
         console.log(`${message} ${data.join(' ')}`);
@@ -44,43 +43,43 @@ export function runTest(testManager: BaseTestManager, testResultsService: ITestR
     // tslint:disable-next-line:no-empty
     server.on('log', (message: string, ...data: string[]) => {
     });
-    // tslint:disable-next-line:no-empty
-    server.on('connect', (data) => {
+    // tslint:disable-next-line:no-empty no-any
+    server.on('connect', (data: any) => {
     });
     // tslint:disable-next-line:no-empty
     server.on('start', (data: { test: string }) => {
     });
     server.on('result', (data: ITestData) => {
-        const test = tests.testFunctions.find(t => t.testFunction.nameToRun === data.test);
-        const statusDetails = outcomeMapping.get(data.outcome);
+        const test = options.tests.testFunctions.find(t => t.testFunction.nameToRun === data.test);
+        const statusDetails = outcomeMapping.get(data.outcome)!;
         if (test) {
             test.testFunction.status = statusDetails.status;
             test.testFunction.message = data.message;
             test.testFunction.traceback = data.traceback;
-            tests.summary[statusDetails.summaryProperty] += 1;
+            options.tests.summary[statusDetails.summaryProperty] += 1;
 
             if (failFast && (statusDetails.summaryProperty === 'failures' || statusDetails.summaryProperty === 'errors')) {
                 testManager.stop();
             }
         } else {
             if (statusDetails) {
-                tests.summary[statusDetails.summaryProperty] += 1;
+                options.tests.summary[statusDetails.summaryProperty] += 1;
             }
         }
     });
-    // tslint:disable-next-line:no-empty
-    server.on('socket.disconnected', (data) => {
+    // tslint:disable-next-line:no-empty no-any
+    server.on('socket.disconnected', (data: any) => {
     });
 
     return server.start().then(port => {
-        const testPaths: string[] = getIdsOfTestsToRun(tests, testsToRun);
+        const testPaths: string[] = getIdsOfTestsToRun(options.tests, options.testsToRun!);
         for (let counter = 0; counter < testPaths.length; counter += 1) {
             testPaths[counter] = `-t${testPaths[counter].trim()}`;
         }
-        const startTestDiscoveryDirectory = getStartDirectory(args);
+        const startTestDiscoveryDirectory = getStartDirectory(options.args);
 
         function runTestInternal(testFile: string = '', testId: string = '') {
-            let testArgs = buildTestArgs(args);
+            let testArgs = buildTestArgs(options.args);
             failFast = testArgs.indexOf('--uf') >= 0;
             testArgs = testArgs.filter(arg => arg !== '--uf');
 
@@ -92,16 +91,25 @@ export function runTest(testManager: BaseTestManager, testResultsService: ITestR
             if (testFile.length > 0) {
                 testArgs.push(`--testFile=${testFile}`);
             }
-            if (debug === true) {
-                return debugLauncher.getPort(Uri.file(rootDirectory))
+            if (options.debug === true) {
+                const debugLauncher = serviceContainer.get<ITestDebugLauncher>(ITestDebugLauncher);
+                return debugLauncher.getPort(options.workspaceFolder)
                     .then(debugPort => {
                         testArgs.push(...['--secret=my_secret', `--port=${debugPort}`]);
+                        const launchOptions = { cwd: options.cwd, args: [testLauncherFile].concat(testArgs), token: options.token, outChannel: options.outChannel, port: debugPort };
                         // tslint:disable-next-line:prefer-type-cast no-any
-                        return debugLauncher.launchDebugger(rootDirectory, [testLauncherFile].concat(testArgs), debugPort, token, outChannel);
+                        return debugLauncher.launchDebugger(launchOptions);
                     });
             } else {
                 // tslint:disable-next-line:prefer-type-cast no-any
-                return run(PythonSettings.getInstance(Uri.file(rootDirectory)).pythonPath, [testLauncherFile].concat(testArgs), rootDirectory, token, outChannel);
+                const runOptions: Options = {
+                    args: [testLauncherFile].concat(testArgs),
+                    cwd: options.cwd,
+                    outChannel: options.outChannel,
+                    token: options.token,
+                    workspaceFolder: options.workspaceFolder
+                };
+                return run(serviceContainer, 'unittest', runOptions);
             }
         }
 
@@ -112,22 +120,22 @@ export function runTest(testManager: BaseTestManager, testResultsService: ITestR
 
         // Ok, the ptvs test runner can only work with one test at a time
         let promise = Promise.resolve<string>('');
-        if (Array.isArray(testsToRun.testFile)) {
-            testsToRun.testFile.forEach(testFile => {
+        if (Array.isArray(options.testsToRun!.testFile)) {
+            options.testsToRun!.testFile!.forEach(testFile => {
                 // tslint:disable-next-line:prefer-type-cast no-any
                 promise = promise.then(() => runTestInternal(testFile.fullPath, testFile.nameToRun) as Promise<any>);
             });
         }
-        if (Array.isArray(testsToRun.testSuite)) {
-            testsToRun.testSuite.forEach(testSuite => {
-                const testFileName = tests.testSuites.find(t => t.testSuite === testSuite).parentTestFile.fullPath;
+        if (Array.isArray(options.testsToRun!.testSuite)) {
+            options.testsToRun!.testSuite!.forEach(testSuite => {
+                const testFileName = options.tests.testSuites.find(t => t.testSuite === testSuite)!.parentTestFile.fullPath;
                 // tslint:disable-next-line:prefer-type-cast no-any
                 promise = promise.then(() => runTestInternal(testFileName, testSuite.nameToRun) as Promise<any>);
             });
         }
-        if (Array.isArray(testsToRun.testFunction)) {
-            testsToRun.testFunction.forEach(testFn => {
-                const testFileName = tests.testFunctions.find(t => t.testFunction === testFn).parentTestFile.fullPath;
+        if (Array.isArray(options.testsToRun!.testFunction)) {
+            options.testsToRun!.testFunction!.forEach(testFn => {
+                const testFileName = options.tests.testFunctions.find(t => t.testFunction === testFn)!.parentTestFile.fullPath;
                 // tslint:disable-next-line:prefer-type-cast no-any
                 promise = promise.then(() => runTestInternal(testFileName, testFn.nameToRun) as Promise<any>);
             });
@@ -135,8 +143,8 @@ export function runTest(testManager: BaseTestManager, testResultsService: ITestR
         // tslint:disable-next-line:prefer-type-cast no-any
         return promise as Promise<any>;
     }).then(() => {
-        testResultsService.updateResults(tests);
-        return tests;
+        testResultsService.updateResults(options.tests);
+        return options.tests;
     }).catch(reason => {
         return Promise.reject(reason);
     });
@@ -186,7 +194,7 @@ function buildTestArgs(args: string[]): string[] {
     return testArgs;
 }
 function getIdsOfTestsToRun(tests: Tests, testsToRun: TestsToRun): string[] {
-    const testIds = [];
+    const testIds: string[] = [];
     if (testsToRun && testsToRun.testFolder) {
         // Get test ids of files in these folders
         testsToRun.testFolder.map(folder => {
