@@ -15,9 +15,9 @@ export type PythonSettingKeys = 'workspaceSymbols.enabled' | 'pythonPath' |
     'unitTest.nosetestArgs' | 'unitTest.pyTestArgs' | 'unitTest.unittestArgs' |
     'formatting.provider' | 'sortImports.args' |
     'unitTest.nosetestsEnabled' | 'unitTest.pyTestEnabled' | 'unitTest.unittestEnabled' |
-    'linting.enabledWithoutWorkspace';
+    'linting.enabledWithoutWorkspace' | 'envFile';
 
-export async function updateSetting(setting: PythonSettingKeys, value: {}, resource: Uri | undefined, configTarget: ConfigurationTarget) {
+export async function updateSetting(setting: PythonSettingKeys, value: {} | undefined, resource: Uri | undefined, configTarget: ConfigurationTarget) {
     const settings = workspace.getConfiguration('python', resource);
     const currentValue = settings.inspect(setting);
     if (currentValue !== undefined && ((configTarget === ConfigurationTarget.Global && currentValue.globalValue === value) ||
@@ -52,7 +52,7 @@ export function retryAsync(wrapped: Function, retryCount: number = 2) {
 
             const makeCall = () => {
                 // tslint:disable-next-line:no-unsafe-any no-any no-invalid-this
-                wrapped.call(this, ...args)
+                wrapped.call(this as Function, ...args)
                     // tslint:disable-next-line:no-unsafe-any no-any
                     .then(resolve, (reason: any) => {
                         reasons.push(reason);
