@@ -1,27 +1,30 @@
 import * as vscode from 'vscode';
 
 export class MockOutputChannel implements vscode.OutputChannel {
+    public name: string;
+    public output: string;
+    public isShown: boolean;
     constructor(name: string) {
         this.name = name;
         this.output = '';
     }
-    name: string;
-    output: string;
-    isShown: boolean;
-    append(value: string) {
+    public append(value: string) {
         this.output += value;
     }
-    appendLine(value: string) { this.append(value); this.append('\n'); }
-    clear() { }
-    show(preservceFocus?: boolean): void;
-    show(column?: vscode.ViewColumn, preserveFocus?: boolean): void;
-    show(x?: any, y?: any): void {
+    public appendLine(value: string) { this.append(value); this.append('\n'); }
+    // tslint:disable-next-line:no-empty
+    public clear() { }
+    public show(preservceFocus?: boolean): void;
+    public show(column?: vscode.ViewColumn, preserveFocus?: boolean): void;
+    // tslint:disable-next-line:no-any
+    public show(x?: any, y?: any): void {
         this.isShown = true;
     }
-    hide() {
+    public hide() {
         this.isShown = false;
     }
-    dispose() { }
+    // tslint:disable-next-line:no-empty
+    public dispose() { }
 }
 
 export class MockStatusBarItem implements vscode.StatusBarItem {
@@ -31,13 +34,13 @@ export class MockStatusBarItem implements vscode.StatusBarItem {
     public tooltip: string;
     public color: string;
     public command: string;
-    show(): void {
-
+    // tslint:disable-next-line:no-empty
+    public show(): void {
     }
-    hide(): void {
-
+    // tslint:disable-next-line:no-empty
+    public hide(): void {
     }
-    dispose(): void {
-
+    // tslint:disable-next-line:no-empty
+    public dispose(): void {
     }
 }
