@@ -4,16 +4,13 @@
 import { expect, use } from 'chai';
 import * as chaiAsPromised from 'chai-as-promised';
 import * as path from 'path';
+import { Uri } from 'vscode';
 import { Product } from '../../client/common/installer';
-import { CANCELLATION_REASON, CommandSource } from '../../client/unittests/common/constants';
-import { TestCollectionStorageService } from '../../client/unittests/common/storageService';
-import { TestResultsService } from '../../client/unittests/common/testResultsService';
-import { TestsHelper } from '../../client/unittests/common/testUtils';
-import { ITestCollectionStorageService, ITestResultsService, ITestsHelper } from '../../client/unittests/common/types';
-import { TestResultDisplay } from '../../client/unittests/display/main';
+import { CANCELLATION_REASON, CommandSource, UNITTEST_PROVIDER } from '../../client/unittests/common/constants';
+import { ITestDiscoveryService } from '../../client/unittests/common/types';
 import { initialize, initializeTest } from '../initialize';
-import { MockOutputChannel } from '../mockClasses';
-import { MockTestManagerWithRunningTests } from './mocks';
+import { MockDiscoveryService, MockTestManagerWithRunningTests } from './mocks';
+import { UnitTestIocContainer } from './serviceRegistry';
 
 use(chaiAsPromised);
 
@@ -35,42 +32,42 @@ const EmptyTests = {
 
 // tslint:disable-next-line:max-func-body-length
 suite('Unit Tests Stopping Discovery and Runner', () => {
-    let testResultDisplay: TestResultDisplay;
-    let outChannel: MockOutputChannel;
-    let storageService: ITestCollectionStorageService;
-    let resultsService: ITestResultsService;
-    let testsHelper: ITestsHelper;
+    let ioc: UnitTestIocContainer;
     suiteSetup(initialize);
     setup(async () => {
-        outChannel = new MockOutputChannel('Python Test Log');
-        testResultDisplay = new TestResultDisplay(outChannel);
         await initializeTest();
+        initializeDI();
     });
-    teardown(() => {
-        outChannel.dispose();
-        testResultDisplay.dispose();
-    });
+    teardown(() => ioc.dispose());
 
-    function createTestManagerDepedencies() {
-        storageService = new TestCollectionStorageService();
-        resultsService = new TestResultsService();
-        testsHelper = new TestsHelper();
+    function initializeDI() {
+        ioc = new UnitTestIocContainer();
+        ioc.registerCommonTypes();
+        ioc.registerProcessTypes();
+        ioc.registerVariableTypes();
+
+        ioc.registerTestParsers();
+        ioc.registerTestVisitors();
+        ioc.registerTestResultsHelper();
+        ioc.registerTestStorage();
+        ioc.registerTestsHelper();
+
     }
 
     test('Running tests should not stop existing discovery', async () => {
-        createTestManagerDepedencies();
-        const mockTestManager = new MockTestManagerWithRunningTests('unittest', Product.unittest, testFilesPath, outChannel, storageService, resultsService, testsHelper);
+        const mockTestManager = new MockTestManagerWithRunningTests(UNITTEST_PROVIDER, Product.unittest, Uri.file(testFilesPath), testFilesPath, ioc.serviceContainer);
+        ioc.serviceManager.addSingletonInstance<ITestDiscoveryService>(ITestDiscoveryService, new MockDiscoveryService(mockTestManager.discoveryDeferred.promise), UNITTEST_PROVIDER);
+
         const discoveryPromise = mockTestManager.discoverTests(CommandSource.auto);
         mockTestManager.discoveryDeferred.resolve(EmptyTests);
-        // tslint:disable-next-line:no-floating-promises
         mockTestManager.runTest(CommandSource.ui);
 
         await expect(discoveryPromise).to.eventually.equal(EmptyTests);
     });
 
     test('Discovering tests should stop running tests', async () => {
-        createTestManagerDepedencies();
-        const mockTestManager = new MockTestManagerWithRunningTests('unittest', Product.unittest, testFilesPath, outChannel, storageService, resultsService, testsHelper);
+        const mockTestManager = new MockTestManagerWithRunningTests(UNITTEST_PROVIDER, Product.unittest, Uri.file(testFilesPath), testFilesPath, ioc.serviceContainer);
+        ioc.serviceManager.addSingletonInstance<ITestDiscoveryService>(ITestDiscoveryService, new MockDiscoveryService(mockTestManager.discoveryDeferred.promise), UNITTEST_PROVIDER);
         mockTestManager.discoveryDeferred.resolve(EmptyTests);
         await mockTestManager.discoverTests(CommandSource.auto);
         const runPromise = mockTestManager.runTest(CommandSource.ui);
@@ -78,7 +75,6 @@ suite('Unit Tests Stopping Discovery and Runner', () => {
         await new Promise(resolve => setTimeout(resolve, 1000));
 
         // User manually discovering tests will kill the existing test runner.
-        // tslint:disable-next-line:no-floating-promises
         mockTestManager.discoverTests(CommandSource.ui, true, false, true);
         await expect(runPromise).to.eventually.be.rejectedWith(CANCELLATION_REASON);
     });
